fix: split mobile workspace tabs panel menu under the line limit
The static max-lines ratchet failed because workspace_tabs_screen.dart grew past 500 lines. Move panel icons, labels, and overflow-menu types into a dedicated part file.
