+++ mobile/lib/src/features/terminal/presentation/workspace_tabs_screen.dart
@@ -29,6 +29,7 @@ import 'package:logging/logging.dart';
 
 part 'workspace_tab_strip.dart';
 part 'workspace_tabs_close.dart';
+part 'workspace_tabs_panel_menu.dart';
 
 /// Tabs of one workspace: a horizontally scrollable chip switcher with one
 /// tab visible at a time. Splits stay a desktop concept.
@@ -452,60 +453,3 @@ class _WorkspaceTabsScreenState extends ConsumerState<WorkspaceTabsScreen> {
     };
   }
 }
-
-IconData _panelIcon(WorkspacePanelDestination destination) {
-  return switch (destination) {
-    WorkspacePanelDestination.terminal => AleraIcons.terminal,
-    WorkspacePanelDestination.explorer => AleraIcons.folder,
-    WorkspacePanelDestination.search => AleraIcons.search,
-    WorkspacePanelDestination.sourceControl => AleraIcons.gitBranch,
-    WorkspacePanelDestination.pullRequest => AleraIcons.gitPullRequest,
-  };
-}
-
-String _panelLabel(WorkspacePanelDestination destination) {
-  return switch (destination) {
-    WorkspacePanelDestination.terminal => 'Terminal',
-    WorkspacePanelDestination.explorer => 'Explorer',
-    WorkspacePanelDestination.search => 'Search',
-    WorkspacePanelDestination.sourceControl => 'Source Control',
-    WorkspacePanelDestination.pullRequest => 'Pull Request',
-  };
-}
-
-class const _PanelMenuRow({
-  required final IconData icon,
-  required final String label,
-  required final bool selected,
-}) extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: <Widget>[
-        Icon(icon, size: 18, color: AleraTokens.foregroundMuted),
-        const SizedBox(width: AleraTokens.space12),
-        Expanded(child: Text(label)),
-        if (selected)
-          const Icon(AleraIcons.check, size: 16, color: AleraTokens.foreground),
-      ],
-    );
-  }
-}
-
-sealed class _NewTabAction {
-  const _NewTabAction();
-}
-
-class const _NewTerminalTabAction() extends _NewTabAction {}
-
-class const _NewAgentProfileTabAction(final String profileId)
-    extends _NewTabAction {}
-
-sealed class _TabsMenuAction {
-  const _TabsMenuAction();
-}
-
-class const _QuickKeysMenuAction() extends _TabsMenuAction {}
-
-class const _SelectPanelAction(final WorkspacePanelDestination destination)
-    extends _TabsMenuAction {}
